.history/requirements_20220619211951.txt
camelot-py
pdfplumber==0.7.1
xlsxwriter
pandas
streamlit